Fix whatsmars misplaced corrections
Previously correct-usage file for
whatsmars misuse 10 & 11 contained
a correct pattern for insecure parameter
to cipher object type of misuse, however
the 10 & 11 yml file lists misuse as the following:
First parameter while initializing SecretKeySpec
was not properly randomized.

Replaced the correct-usages files
with the same solution as for whatsmars
misuse 6, because these are the same misuse issue.